fix(backend): align publication recovery schema
